% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_patents.R
\name{read_patents}
\alias{read_patents}
\title{Reading US Patents}
\usage{
read_patents(file_dir)
}
\arguments{
\item{file_dir}{A folder of csv files from the US Patent Database}
}
\value{
A data frame of binded patents
}
\description{
Reading in patent files and binding them
}
